+++ tests/actor/test_command_auto_pilot.py
@@ -178,3 +178,47 @@ async def test_command_auto_pilot_add_then_removehartmann_while_running(
             "guide",
         ],
     )
+
+
+async def test_command_auto_pilot_goto_running(
+    actor: HALActor,
+    mock_auto_pilot,
+    mocker: MockerFixture,
+):
+    goto = actor.helpers.macros["goto_field"]
+    mocker.patch.object(goto, "_running", True)
+
+    wait_until_complete_mock = mocker.patch.object(
+        goto,
+        "wait_until_complete",
+        return_value=True,
+    )
+
+    cmd = actor.invoke_mock_command("auto-pilot")
+    await cmd
+
+    assert cmd.status.did_succeed
+
+    wait_until_complete_mock.assert_called()
+
+
+async def test_command_auto_pilot_goto_running_fails(
+    actor: HALActor,
+    mock_auto_pilot,
+    mocker: MockerFixture,
+):
+    goto = actor.helpers.macros["goto_field"]
+    mocker.patch.object(goto, "_running", True)
+
+    wait_until_complete_mock = mocker.patch.object(
+        goto,
+        "wait_until_complete",
+        return_value=False,
+    )
+
+    cmd = actor.invoke_mock_command("auto-pilot")
+    await cmd
+
+    assert cmd.status.did_fail
+
+    wait_until_complete_mock.assert_called()